requirements.txt
aws-cdk-lib==2.2.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecs_patterns
aws_cdk.aws_applicationautoscaling
aws_cdk.aws_iam
aws_cdk.aws_ssm
